renamed utils.ts file and relocated it to utils folder
