+++ scripts/utils/convertToJson.ts
